pyproject.toml
[tool.poetry]
name = "mqtt-processor"
version = "0.1.0"
description = "Customizable application that receives MQTT messages, filters or changes the message and re-sends it to another topic\""
authors = ["Ondřej Novák <onno.cz@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.9.0"
PyYAML = "^6.0"
paho-mqtt = "^1.6.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
